+++ bedwars-plugin/pom.xml
@@ -184,7 +184,7 @@
         </dependency>
         <dependency>
             <groupId>com.andrei1058.bedwars</groupId>
-            <artifactId>versionsupport_v1_20_R1</artifactId>
+            <artifactId>versionsupport_v1_20_R2</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
