+++ pom.xml
@@ -111,7 +111,7 @@
         <module>versionsupport_v1_18_R2</module>
         <module>versionsupport_v1_19_R2</module>
         <module>versionsupport_v1_19_R3</module>
-        <module>versionsupport_v1_20_R1</module>
+        <module>versionsupport_v1_20_R2</module>
     </modules>
 
     <distributionManagement>
