--- versionsupport_v1_20_R1/src/main/java/com/andrei1058/bedwars/support/version/v1_20_R1/despawnable/DespawnableType.java
@@ -1,6 +1,0 @@
-package com.andrei1058.bedwars.support.version.v1_20_R1.despawnable;
-
-public enum DespawnableType {
-    IRON_GOLEM,
-    SILVERFISH
-}
