+++ versionsupport_v1_20_R2/src/main/java/com/andrei1058/bedwars/support/version/v1_20_R2/despawnable/DespawnableAttributes.java
@@ -1,4 +1,4 @@
-package com.andrei1058.bedwars.support.version.v1_20_R1.despawnable;
+package com.andrei1058.bedwars.support.version.v1_20_R2.despawnable;
 
 public record DespawnableAttributes(DespawnableType type, double speed, double health, double damage, int despawnSeconds) {
